remove foobar dto from lib/ (moved to examples/)
