+++ examples/CMakeLists.txt
@@ -3,6 +3,7 @@ cmake_policy(VERSION 3.13)
 
 etn_target(exe ${PROJECT_NAME}-example
     SOURCES
+        src/foobar_dto.cc
         src/fty_common_messagebus_example.cc
     USES
         ${PROJECT_NAME_UNDERSCORE}
@@ -11,6 +12,7 @@ etn_target(exe ${PROJECT_NAME}-example
 
 etn_target(exe ${PROJECT_NAME}-example2
     SOURCES
+        src/foobar_dto.cc
         src/fty_common_messagebus_example2.cc
     USES
         ${PROJECT_NAME_UNDERSCORE}
@@ -19,6 +21,7 @@ etn_target(exe ${PROJECT_NAME}-example2
 
 etn_target(exe ${PROJECT_NAME}-example_req
     SOURCES
+        src/foobar_dto.cc
         src/fty_common_messagebus_example_req.cc
     USES
         ${PROJECT_NAME_UNDERSCORE}
@@ -27,6 +30,7 @@ etn_target(exe ${PROJECT_NAME}-example_req
 
 etn_target(exe ${PROJECT_NAME}-example_rep
     SOURCES
+        src/foobar_dto.cc
         src/fty_common_messagebus_example_rep.cc
     USES
         ${PROJECT_NAME_UNDERSCORE}
@@ -35,6 +39,7 @@ etn_target(exe ${PROJECT_NAME}-example_rep
 
 etn_target(exe ${PROJECT_NAME}-example_rep2
     SOURCES
+        src/foobar_dto.cc
         src/fty_common_messagebus_example_rep2.cc
     USES
         ${PROJECT_NAME_UNDERSCORE}
