+++ examples/src/foobar_dto.cc
