+++ examples/src/foobar_dto.h
